# backend/tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentRunCancelRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "default": "user_abort",
            "maxLength": 64,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "AgentRunCancelRequest",
        "type": "object"
      },
      "AgentRunCancelResponse": {
        "additionalProperties": false,
        "properties": {
          "cancel_requested": {
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "run_id",
          "status",
          "cancel_requested"
        ],
        "title": "AgentRunCancelResponse",
        "type": "object"
      },
      "AgentRunCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "goal_type_override": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GoalType"
              },
              {
                "type": "null"
              }
            ]
          },
          "hint_intent": {
            "anyOf": [
              {
                "enum": [
                  "create_plan",
                  "replan"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint Intent"
          },
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "source_plan_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Plan Id"
          }
        },
        "required": [
          "message"
        ],
        "title": "AgentRunCreateRequest",
        "type": "object"
      },
      "AgentRunCreatedResponse": {
        "additionalProperties": false,
        "properties": {
          "events_url": {
            "title": "Events Url",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "run_id",
          "status",
          "events_url"
        ],
        "title": "AgentRunCreatedResponse",
        "type": "object"
      },
      "AgentRunResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "final_plan_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Plan Id"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "replan_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReplanMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolved_intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunIntent"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanResultSummary"
              },
              {
                "$ref": "#/components/schemas/ClarificationRequest"
              },
              {
                "$ref": "#/components/schemas/SafeResponse"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "result_kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunResultKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Category"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "total_cost_cny": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Cost Cny",
            "type": "string"
          },
          "total_latency_ms": {
            "minimum": 0.0,
            "title": "Total Latency Ms",
            "type": "integer"
          },
          "total_tokens_in": {
            "minimum": 0.0,
            "title": "Total Tokens In",
            "type": "integer"
          },
          "total_tokens_out": {
            "minimum": 0.0,
            "title": "Total Tokens Out",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "status",
          "resolved_intent",
          "replan_mode",
          "result_kind",
          "result",
          "final_plan_id",
          "fallback_reason",
          "error_code",
          "risk_category",
          "total_tokens_in",
          "total_tokens_out",
          "total_cost_cny",
          "total_latency_ms",
          "created_at",
          "finished_at"
        ],
        "title": "AgentRunResponse",
        "type": "object"
      },
...
      "DevRunDetail": {
        "additionalProperties": false,
        "properties": {
          "config_snapshot": {
            "$ref": "#/components/schemas/DevSnapshot"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/DevEventTrace"
            },
            "title": "Events",
            "type": "array"
          },
          "input_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DevSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_text": {
            "title": "Request Text",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "run": {
            "$ref": "#/components/schemas/DevRunSummary"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/DevStepTrace"
            },
            "title": "Steps",
            "type": "array"
          },
          "terminal_invariant": {
            "$ref": "#/components/schemas/TerminalInvariant"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/DevToolTrace"
            },
            "title": "Tools",
            "type": "array"
          }
        },
        "required": [
          "run",
          "request_text",
          "input_snapshot",
          "config_snapshot",
          "result",
          "steps",
          "tools",
          "events",
          "terminal_invariant"
        ],
        "title": "DevRunDetail",
        "type": "object"
      },
      "DevRunListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DevRunSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "DevRunListResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Stable API error envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "GoalType": {
        "enum": [
          "ai_backend",
          "agent_app",
          "backend_java",
          "data_engineer",
          "fullstack",
          "other"
        ],
        "title": "GoalType",
        "type": "string"
      },
      "GuestLoginRequest": {
        "additionalProperties": false,
        "description": "Optional stable browser device identifier for Guest reuse.",
        "properties": {
          "device_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          }
        },
        "title": "GuestLoginRequest",
        "type": "object"
      },
      "GuestLoginResponse": {
        "additionalProperties": false,
        "description": "Bearer token returned by Guest authentication.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "minimum": 1.0,
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserSummary"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "GuestLoginResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonValue": {},
      "MeResponse": {
        "additionalProperties": false,
        "description": "Stage 1 home restoration aggregate.",
        "properties": {
          "active_plan": {
            "title": "Active Plan",
            "type": "null"
          },
          "active_run": {
            "title": "Active Run",
            "type": "null"
          },
          "latest_review": {
            "title": "Latest Review",
            "type": "null"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfileResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "profile_complete": {
            "title": "Profile Complete",
            "type": "boolean"
          },
          "today_tasks": {
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Today Tasks",
            "type": "array"
          },
          "user": {
            "$ref": "#/components/schemas/UserSummary"
          }
        },
        "required": [
          "user",
          "profile_complete",
          "profile"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "ReplanMode": {
        "enum": [
          "initial",
          "continue",
          "adjust"
        ],
        "title": "ReplanMode",
        "type": "string"
      },
...
      "RunIntent": {
        "enum": [
          "create_plan",
          "replan",
          "unsupported"
        ],
        "title": "RunIntent",
        "type": "string"
      },
      "RunResultKind": {
        "enum": [
          "plan",
          "clarification",
          "safe_response"
        ],
        "title": "RunResultKind",
        "type": "string"
      },
      "RunStatus": {
        "enum": [
          "pending",
          "running",
          "completed",
          "degraded",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Career Planning Buddy",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/agent-runs": {
      "post": {
        "operationId": "create_agent_run_api_v1_agent_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Agent Run",
        "tags": [
          "agent-runs"
        ]
      }
    },
    "/api/v1/agent-runs/{run_id}": {
      "get": {
        "operationId": "get_agent_run_api_v1_agent_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Agent Run",
        "tags": [
          "agent-runs"
        ]
      }
    },
    "/api/v1/agent-runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_agent_run_api_v1_agent_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunCancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Agent Run",
        "tags": [
          "agent-runs"
        ]
      }
    },
    "/api/v1/agent-runs/{run_id}/events": {
      "get": {
        "operationId": "stream_agent_run_events_api_v1_agent_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stream Agent Run Events",
        "tags": [
          "agent-runs"
        ]
      }
    },
    "/api/v1/auth/guest": {
      "post": {
        "operationId": "guest_login_api_v1_auth_guest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestLoginResponse"
                }
              }
            },
            "description": "Existing Guest reused"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestLoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "summary": "Guest Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/dev/runs": {
      "get": {
        "operationId": "list_runs_api_v1_dev_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "result_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result Kind"
            }
          },
          {
            "in": "query",
            "name": "error_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error Code"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Runs",
        "tags": [
          "developer"
        ]
      }
    },
    "/api/v1/dev/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_dev_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevRunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run",
        "tags": [
          "developer"
        ]
      }
    },
...
    "/api/v1/me": {
      "get": {
        "operationId": "get_me_api_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid token"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}